+++ app/src/main/java/com/rodolfo/cashflow/domain/models/Transferencia.java
@@ -15,8 +15,8 @@ import lombok.ToString;
 public class Transferencia extends Transaccion{
 
     public Transferencia (BigDecimal monto,
-                          Integer usuario_origen,
-                          Integer usuario_destino,
+                          Long usuario_origen,
+                          Long usuario_destino,
                           String descripcion) {
 
         super(monto, "TRANSFERENCIA", usuario_origen, usuario_destino, null, descripcion);
